refactor(devtools): migrate panel to separate file
